Update patch set 1
Patch Set 1:

This is still needed. This changes the pressure range reported by the touchscreen. Our values are very close to zero. This scales them up for applications such as Markers and other drawing programs that use the pressure variation to determine line width. I think the touch peak and touch detect thresholds just determine whether a touch has ocured. In markers, the lines are always about the same width without this.This scales touches to a usable range.

Patch-set: 1