Update patch set 1
Patch Set 1:

I have to run out soon, so I'm letting VeryGreen run some builds on this.  If he doesn't get to it in a few hours, I'll run them and approve.

FYI, the original source for this is:
https://github.com/omxcodec/android_frameworks_native/commit/05ac26422086b1c357d9dff456a6f037120f42bc

So on my review patch I set author= to him for reference.

Patch-set: 1